pyproject.toml
[project]
name = "indian-screener"
version = "1.0.0"
description = "Production-grade financial database and API replicating screener.in's data layer"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy>=2.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9",
    # Scraping
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "playwright>=1.48",
    # Data processing
    "pandas>=2.2",
    "numpy>=1.26",
    # Scheduling
    "apscheduler>=3.10",
    # Validation & config
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    # Utilities
    "tenacity>=9.0",
    "python-dateutil>=2.9",
    # Redis for caching
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
